Updated TableManager to DeletionManager
The previous TableManager class has been converted to a DeletionManager with an interface to match. This is part of a redesign of the sub-managers so that there is a sub manager class for each of the CRUD operations. This class will deal with all operations surrounding deletion in the database.