Updated CreationManager and Added CommandBuilder
CreationManager now uses the new CommandBuilder to get a list of all the commands it needs to run. This is to make implementation of new classes (like UpdateManager) and methods easier and it reduces repeat code.

The CommandBuilder uses a lot of what was in CreationManager previously but reworks it to return a list of all commands needed.